+++ .gitignore
@@ -37,3 +37,6 @@ out/
 ### VS Code ###
 .vscode/
 /.apt_generated_tests/
+
+### application.properties ###
+.application.properties
