+++ build.gradle
@@ -17,16 +17,42 @@ repositories {
     maven { url 'https://repo.spring.io/snapshot' }
 }
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation group: 'io.jsonwebtoken', name: 'jjwt', version: '0.9.1'
-    implementation 'org.springframework.boot:spring-boot-starter-mail'
-    compileOnly 'org.projectlombok:lombok'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-    annotationProcessor 'org.projectlombok:lombok'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    // 스프링 데이터 JPA
+    implementation('org.springframework.boot:spring-boot-starter-data-jpa')
+    // 스프링부트 웹 관련 기능 제공 | 테스트 기능을 제공
+    implementation('org.springframework.boot:spring-boot-starter-web')
+    // 스프링 시큐리티 사용 스타터
+    implementation('org.springframework.boot:spring-boot-starter-security')
+    // 프로젝트 유효성 검사를 위한 의존성 입력
+    implementation('jakarta.validation:jakarta.validation-api:3.0.2')
+    implementation('commons-validator:commons-validator:1.7')
+
+    // 스프링 잉메일 기능 제공
+    implementation('org.springframework.boot:spring-boot-starter-mail')
+    implementation('org.springframework.boot:spring-boot-starter-web')
+
+    // JWT 의존성 추가
+    implementation('io.jsonwebtoken:jjwt-api:0.11.2')
+    implementation('io.jsonwebtoken:jjwt-impl:0.11.2')
+    implementation('io.jsonwebtoken:jjwt-jackson:0.11.2')
+
+    // 데이터베이스 h2 사용 - DB 테이블을 자바 객체로 가져오는 용도
+    runtimeOnly('com.h2database:h2')
+    implementation('mysql:mysql-connector-java:8.0.33')
+
+    // 반복 메서드 작성 작업을 줄여주는 라이브러리
+    compileOnly('org.projectlombok:lombok')
+    annotationProcessor('org.projectlombok:lombok')
+
+    testImplementation('org.junit.jupiter:junit-jupiter:5.8.2')
+    testImplementation('org.assertj:assertj-core:3.22.0')
+    testImplementation('org.springframework.boot:spring-boot-starter-test')
+    // 스프링 시큐리티 테스트를 위한 테스트 기능 제공
+    testImplementation('org.springframework.security:spring-security-test')
+
+    // 프로젝트 유효성 검사를 위한 의존성 입력
+    implementation('jakarta.validation:jakarta.validation-api:3.0.2')
+    implementation('commons-validator:commons-validator:1.7')
 }
 tasks.named('test') {
     useJUnitPlatform()
