+++ src/main/java/com/mong/mmbs/controller/ProductCategoryController.java
@@ -11,13 +11,14 @@ import com.mong.mmbs.dto.response.ResponseDto;
 import com.mong.mmbs.dto.response.productCategory.ProductAgeGetListResponseDto;
 import com.mong.mmbs.dto.response.productCategory.ProductGenreGetListResponseDto;
 import com.mong.mmbs.dto.response.productCategory.ProductGetResponseDto;
-import com.mong.mmbs.service.ProductCategoryService;
+import com.mong.mmbs.service.implement.ProductCategoryServiceImpl;
 
 @RestController
 @RequestMapping(ApiMappingPattern.PRODUCT_CATEGORY)
 public class ProductCategoryController<BookListServicevice> {
 	
-	@Autowired ProductCategoryService productCategoryService;
+	@Autowired
+	ProductCategoryServiceImpl productCategoryService;
 
 	private static final String GET_PRODUCT_AGE_LIST = "/bookList1/{produtAgeCategory}/{productSubAgeCategory}";
 	private static final String GET_PRODUCT_GENRE_LIST = "/bookList2/{produtGenreCategory}/{productSubGenreCategory}";
