+++ src/main/java/com/mong/mmbs/dto/MailDto.java
@@ -4,10 +4,10 @@ import org.springframework.mail.SimpleMailMessage;
 
 public class MailDto extends SimpleMailMessage {
 
-    private static final String SUBJECT = "어린이 서점@@@ 임시 비밀번호 안내 이메일입니다.";
-    private static final String TEXT = "안녕하세요. 어린이 서점@@@ 임시 비밀번호 안내 메일입니다. "
+    private static final String SUBJECT = "임시 비밀번호 안내 이메일입니다.";
+    private static final String TEXT = "안녕하세요. 임시 비밀번호 안내 메일입니다. "
         + "\n" + "회원님의 임시 비밀번호는 아래와 같습니다. 로그인 후 반드시 비밀번호를 변경해주세요." + "\n";
-    private static final String FROM = "helpringproject@gmail.com";
+    private static final String FROM = "helpringproject@gmail.com"; // 발신자 이메일
 
     public MailDto(String tmpPassword, String memberEmail) {
