+++ src/main/java/com/mong/mmbs/dto/request/ask/AskPostRequestDto.java
@@ -1,8 +1,8 @@
 package com.mong.mmbs.dto.request.ask;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 import lombok.NoArgsConstructor;
