+++ src/main/java/com/mong/mmbs/dto/request/auth/FindIdRequestDto.java
@@ -1,7 +1,6 @@
 package com.mong.mmbs.dto.request.auth;
 
-import javax.validation.constraints.NotBlank;
-
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 import lombok.NoArgsConstructor;
