+++ src/main/java/com/mong/mmbs/dto/request/auth/SignInRequestDto.java
@@ -1,14 +1,13 @@
 package com.mong.mmbs.dto.request.auth;
 
-import javax.validation.constraints.NotBlank;
-
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
 public class SignInRequestDto {
-  @NotBlank
+	@NotBlank
 	private String userId;
 	@NotBlank
 	private String userPassword;
