+++ src/main/java/com/mong/mmbs/dto/request/cart/CartPostRequestDto.java
@@ -1,7 +1,6 @@
 package com.mong.mmbs.dto.request.cart;
 
-import javax.validation.constraints.NotNull;
-
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 import lombok.NoArgsConstructor;
