+++ src/main/java/com/mong/mmbs/dto/request/order/OrderPostRequestDto.java
@@ -1,8 +1,7 @@
 package com.mong.mmbs.dto.request.order;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -36,8 +35,8 @@ public class OrderPostRequestDto {
 
 	@NotNull
 	private int productId;
-  @NotNull
-  private int orderCount;
+  	@NotNull
+  	private int orderCount;
 
 	public boolean getOrderUserWhether() {
 		return this.orderUserWhether;
