+++ src/main/java/com/mong/mmbs/dto/request/product/ProductLikePostRequestDto.java
@@ -1,7 +1,6 @@
 package com.mong.mmbs.dto.request.product;
 
-import javax.validation.constraints.NotNull;
-
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,6 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ProductLikePostRequestDto {
+
 	@NotNull
 	private int productSeq;
 }
