+++ src/main/java/com/mong/mmbs/entity/AskEntity.java
@@ -8,11 +8,11 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import javax.validation.constraints.NotBlank;
 
 import com.mong.mmbs.dto.request.ask.AskPatchRequestDto;
 import com.mong.mmbs.dto.request.ask.AskPostRequestDto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
