+++ src/main/resources/application.properties
@@ -1,8 +1,11 @@
-server.port=4080
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+server.port=4040
+
 spring.datasource.url=jdbc:mysql://localhost:3306
 spring.datasource.username=root
 spring.datasource.password=root
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=true
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
@@ -10,3 +13,12 @@ spring.mail.username=
 spring.mail.password=root
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
+
+spring.mail.properties.mail.smtp.connectiontimeout=5000
+spring.mail.properties.mail.smtp.timeout=5000
+spring.mail.properties.mail.smtp.writetimeout=5000
+
+mail.auth-code-expiration-millis=180000
+
+jwt.secret=1234567890qwertyuiopasdfghjklzxcvbnmQWERTYUIOPASDFGHJKLZXCVBNM1234567890
+jwt.expiration=36000000
